app/requirements.txt
# This file is used by pip to install required python packages
# Usage: pip install -r requirements.txt

Flask
Blueprint
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
Flask_Restful
requests
python-dateutil
